% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funds.R
\name{categorize_op_funds}
\alias{categorize_op_funds}
\title{Categorize operating funds}
\usage{
categorize_op_funds(
  df,
  fund_col = "Fund Name",
  cat_col = "Fund Category",
  after = NULL
)
}
\arguments{
\item{df}{A dataframe, has a fund name column.}

\item{fund_col}{A string, name of the column to with fund names to order, defaults
to "Fund Category".}

\item{cat_col}{A string, name to assign to the column of fund categories to be appended. Defaults
to "Fund Category".}

\item{after}{A string, name of a fund to tack on at the end of the factor (usually "Total").}
}
\value{
A dataframe with col converted to factor.
}
\description{
Appends a fund category column based on an existent fund name column.
}
\author{
Lillian Nguyen
}
