% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_agencies.R
\name{combine_agencies}
\alias{combine_agencies}
\title{Combine agency names and IDs}
\usage{
combine_agencies(df)
}
\arguments{
\item{df}{dataframe with agency information that needs to be combined; use
rename_cols() first}
}
\value{
df with modified agency columns
}
\description{
Combine agency data where legacy systems differ from existing municipal organization
}
\seealso{
\code{\link{rename_colsl}}
}
\author{
Lillian Nguyen
}
