% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analyst_dirs.R
\name{create_analyst_dirs}
\alias{create_analyst_dirs}
\title{Create a folder for each BBMR analyst}
\usage{
create_analyst_dirs(path, additional_folders = NULL)
}
\arguments{
\item{path}{A string, the path where the folders should be created}

\item{additional_folders}{A string, or vector of strings, containing the names of additional folders to be created}
}
\description{
Useful for distributing files by analyst. Analyst names are from the .Renviron file.
}
\author{
Lillian Nguyen
}
