% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workday_crosswalk.R
\name{exp_actuals_workday_xwalk}
\alias{exp_actuals_workday_xwalk}
\title{Make actual expenditures crosswalk}
\usage{
exp_actuals_workday_xwalk(
  actuals_file =
    "G:/Fiscal Years/Fiscal 2022/Projections Year/2. Monthly Expenditure Data/Month 12_June Projections/Expenditure 2022-06_Run7.xlsx"
)
}
\arguments{
\item{actuals_file, }{BAPS excel file of monthly expenditure data}
}
\value{
dataframe of result, data frame of missing cost centers
}
\description{
Make actual expenditures crosswalk
}
\author{
Sara Brumfield
}
