% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workday_crosswalk.R
\name{exp_lines_workday_xwalk}
\alias{exp_lines_workday_xwalk}
\title{Make expenditure lines crosswalk}
\usage{
exp_lines_workday_xwalk(exp_file, bpfs_table = "PLANNINGYEAR24")
}
\arguments{
\item{exp_file, }{BPFS-formatted excel planning year file}

\item{bpfs_table, }{"PLANNINGYEAR24" set as default}
}
\value{
dataframe of result, dataframe of missing cost centers
}
\description{
Make expenditure lines crosswalk
}
\author{
Sara Brumfield
}
