% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel_tabs.R
\name{export_excel_tabs}
\alias{export_excel_tabs}
\title{Splits dataframe into Excel sheets by column}
\usage{
export_excel_tabs(df, tab_name, file_name, drop_col = TRUE, col_width = "auto")
}
\arguments{
\item{df}{A data frame, will be exported as an Excel table}

\item{tab_name}{a string, the name of the column which should be used to separate the df into tabs}

\item{file_name}{a string, the name of the file to be created}

\item{drop_col}{optional (defaults to TRUE) logical, if FALSE then the column specified in the tab_name argument remains is not dropped}

\item{col_width}{optional (defaults to 'auto'), either 'auto' for autosized columns or a vector of numbers which must be equal to the number of columns}
}
\value{
A new .xlsx file with records from a single dataframe placed on multiple tabs
}
\description{
Takes one dataframe and exports it as multiple Excel sheets in the same file, with table formatting, headers, auto-column width, etc., based on the categories in a column. Currently does not auto-size columns with dates correctly. This function can be used first, and then ExportExcel(), with the "existing" argument,  used second in order to add a separate sheet.
}
\examples{
export_excel_tabs(iris, "Species", "iris.xlsx")

}
\seealso{
\code{\link{export_excel}}
}
\author{
Jeremy Pesner, Lillian Nguyen
}
