% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formstack.R
\name{extract_fs_submissions}
\alias{extract_fs_submissions}
\title{Extract Formstack submissions}
\usage{
extract_fs_submissions(api, form_id, approved_only = FALSE)
}
\arguments{
\item{api, }{A list containing the API connection to Formstack,
generated from  connect_fs_api()}

\item{form_id}{numeric, Formstack ID for form}

\item{approved_only}{logical, TRUE to filter for only approved submissions}
}
\value{
A df with submission data
}
\description{
Pulls and extracts all data from a Formstack form, regardless of length. Only
keeps submission data; removes all payload data.
}
\author{
Lillian Nguyen
}
