% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workday_get_appropriation_files.R
\name{get_appropriation_files}
\alias{get_appropriation_files}
\title{Retrieves Workday values from past appropriationfiles
Matches Workday fund = fund, spend category = OSO, cost center = program + activity, grant = detailed fund, ledger = payroll natural}
\usage{
get_appropriation_files()
}
\value{
list of dataframes
}
\description{
Retrieves Workday values from past appropriationfiles
Matches Workday fund = fund, spend category = OSO, cost center = program + activity, grant = detailed fund, ledger = payroll natural
}
\author{
Sara Brumfield
}
