% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formstack.R
\name{get_fs_submissions}
\alias{get_fs_submissions}
\title{Get Formstack submissions}
\usage{
get_fs_submissions(api, form_id, query)
}
\arguments{
\item{api, }{A list containing the API connection to Formstack,
generated from  connect_fs_api()}

\item{form_id}{numeric, Formstack ID for form}

\item{query}{list, conforming to Formstack API}

\item{per_page}{Numeric, defaults to the max of 100. The number of
submissions to retrieve per page.}
}
\value{
A list with payload and submission data
}
\description{
Use this for finer-grained control of Formstack submission retrieval. This
function is the backbone of extract_fs_submissions().
}
\author{
Lillian Nguyen
}
