% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fy.R
\name{get_fy}
\alias{get_fy}
\title{Get Fiscal Year from date}
\usage{
get_fy(date = NULL)
}
\arguments{
\item{date}{lubridate object}
}
\value{
numeric
}
\description{
Get Fiscal Year from date
}
\author{
Lillian Nguyen
}
