% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_mod.R
\name{get_last_mod}
\alias{get_last_mod}
\title{Get Name of Last Modified File}
\usage{
get_last_mod(path, pattern, ignore.case = FALSE, recursive = FALSE)
}
\arguments{
\item{path}{string, file path}

\item{pattern}{string, regex}

\item{ignore.case}{logical}

\item{recursive}{logical}
}
\value{
a string with the filename of the last modified file
}
\description{
Get Name of Last Modified File
}
\author{
Lillian Nguyen
}
