% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workday_master_join.R
\name{make_master_join}
\alias{make_master_join}
\title{Creates master dataframe from joined values from Workday, BPFS and past appropriation files
Matches Workday fund = fund, spend category = OSO, cost center = program + activity, grant = detailed fund, ledger = payroll natural}
\usage{
make_master_join()
}
\value{
dataframes of joined values from Workday, BPFS and past appropriation files
}
\description{
Creates master dataframe from joined values from Workday, BPFS and past appropriation files
Matches Workday fund = fund, spend category = OSO, cost center = program + activity, grant = detailed fund, ledger = payroll natural
}
\author{
Sara Brumfield
}
