% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workday_get_values.R
\name{make_program_activity}
\alias{make_program_activity}
\title{Concatenates BPFS values to create a Workday-compatible Program-Activity field}
\usage{
make_program_activity(
  data = df,
  program_col = `Program ID`,
  activity_col = `Activity ID.y`,
  subactivity_col = SUBACTIVITY_ID
)
}
\value{
concatentated string value
}
\description{
Concatenates BPFS values to create a Workday-compatible Program-Activity field
}
\author{
Sara Brumfield
}
