% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_name_to_id.R
\name{match_name_to_id}
\alias{match_name_to_id}
\title{Match Name to ID}
\usage{
match_name_to_id(df, cols, incl.higher, fy = 22)
}
\arguments{
\item{df}{dataframe to match}

\item{cols}{a vector of strings of the column names to be matched}

\item{incl.higher}{logical, should the returned dataframe include the level of detail above certain columns (ex: should Fund ID be included with a Detailed Fund match?)}

\item{fy}{latest fiscal year, for database schema}
}
\value{
a dataframe with Name columns added
}
\description{
Query BPFS database for latest Detailed Fund, Service,
Activity, Fund, Agency, Object, and/or Subobject names.
Useful for pulling the most recent names when working with historical data.
}
\author{
Lillian Nguyen
}
