% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db.R
\name{query_db}
\alias{query_db}
\title{Query Finance BPFS db}
\usage{
query_db(schema, table)
}
\arguments{
\item{schema, }{a character with "planningyearXX"}

\item{table, }{table name}
}
\value{
A connection, still needs to be collected
}
\description{
Queries from the production database
}
\author{
Lillian Nguyen
}
