% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cols.R
\name{rename_cols}
\alias{rename_cols}
\title{Renames commonly inconsistent column names in BBMR data}
\usage{
rename_cols(df)
}
\arguments{
\item{df}{dataframe with colnames to modify}
}
\value{
A vector of strings
}
\description{
Standardizes column names to Title Case when recognized,
and tolowercasewithout spaces when not recognized
}
\author{
Lillian Nguyen
}
