% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workday_crosswalk.R
\name{rev_projections_workday_xwalk}
\alias{rev_projections_workday_xwalk}
\title{Make projections crosswalk}
\usage{
rev_projections_workday_xwalk(fiscal_year = 2023)
}
\arguments{
\item{fiscal_year}{the column to check numbers on, e.g., Fiscal 2023 Budget}
}
\value{
dataframe, numbers check result as text
}
\description{
Make projections crosswalk
}
\author{
Sara Brumfield
}
