% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workday_crosswalk.R
\name{revenue_workday_xwalk}
\alias{revenue_workday_xwalk}
\title{Crosswalks Workday and BPFS values}
\usage{
revenue_workday_xwalk(
  data = data$revenue,
  coa_file = "G:/Analyst Folders/Sara Brumfield/_ref/Chart_of_Accounts 2022-09-06.xlsx",
  coa_sheet = "All Revenue Categories (Non-Gr)"
)
}
\arguments{
\item{data, }{dataframe, default to data$revenue for data prep}

\item{coa_file, }{file path}

\item{coa_sheet}{sheet name for chart of accounts file}
}
\value{
dataframe
}
\description{
Joins datasets on common fields
}
\author{
Sara Brumfield
}
