% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workday_get_values.R
\name{workday_get_values}
\alias{workday_get_values}
\title{Retrieves Workday values from the foundational data model file
Matches Workday fund = fund, spend category = OSO, cost center = program + activity, grant = detailed fund, ledger = payroll natural}
\usage{
workday_get_values()
}
\value{
list of dataframes
}
\description{
Retrieves Workday values from the foundational data model file
Matches Workday fund = fund, spend category = OSO, cost center = program + activity, grant = detailed fund, ledger = payroll natural
}
\author{
Sara Brumfield
}
