% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funds.R
\name{order_funds}
\alias{order_funds}
\title{Order funds}
\usage{
order_funds(df, col = "Fund Name", before = NULL, after = NULL)
}
\arguments{
\item{df}{Dataframe with fund column to modify.}

\item{col}{String. Name of the column to with fund names to order, defaults
to "Fund Name".}

\item{before}{A string, name of a fund to tack on at the beginning of the factor (usually "Total").}

\item{after}{A string, name of a fund to tack on at the end of the factor (usually "Total").}
}
\value{
A dataframe with col converted to factor.
}
\description{
Converts a column of fund name strings to factor in BBMR standardized order
}
\author{
Lillian Nguyen
}
